vista del archivo
parte visual de la prueba bd

+++ indra/bd/multiplicacion.sql
@@ -1,11 +1,11 @@
 -- phpMyAdmin SQL Dump
--- version 5.1.1
+-- version 5.0.4
 -- https://www.phpmyadmin.net/
 --
 -- Servidor: 127.0.0.1
--- Tiempo de generación: 26-01-2022 a las 16:05:45
--- Versión del servidor: 10.4.22-MariaDB
--- Versión de PHP: 7.3.33
+-- Tiempo de generación: 26-01-2022 a las 19:55:23
+-- Versión del servidor: 10.4.17-MariaDB
+-- Versión de PHP: 7.3.26
 
 SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
 START TRANSACTION;
@@ -18,7 +18,7 @@ SET time_zone = "+00:00";
 /*!40101 SET NAMES utf8mb4 */;
 
 --
--- Base de datos: `db_incraprueba`
+-- Base de datos: `db_prueba_practica`
 --
 
 -- --------------------------------------------------------
@@ -29,9 +29,9 @@ SET time_zone = "+00:00";
 
 CREATE TABLE `multiplicacion` (
   `id` int(11) NOT NULL,
-  `num1` bigint(20) NOT NULL,
-  `num2` bigint(20) NOT NULL,
-  `res` bigint(20) NOT NULL,
+  `num1` decimal(65,0) NOT NULL,
+  `num2` decimal(65,0) NOT NULL,
+  `res` decimal(65,0) NOT NULL,
   `fecha` date NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
